feat: move tabs constants to utils
