+++ app/layout/scenarios/edit/tabs/constants.ts
@@ -1,4 +1,4 @@
-import { ScenarioSidebarTabs, ScenarioSidebarSubTabs } from 'layout/scenarios/edit/sidebar/types';
+import { ScenarioSidebarTabs, ScenarioSidebarSubTabs } from 'utils/tabs';
 
 export const TABS = [
   {
